fix(hooks): reject custom Mise shim dirs and seal tool selector (#3105)
