+++ scripts/tests/dev2merge-quality-gate-receipt-tests.sh
@@ -154,6 +154,7 @@ cp "$repo_root/scripts/cargo-env-normalize.sh" "$fixture/scripts/"
 cp "$repo_root/scripts/quality-gate-state.py" "$fixture/scripts/"
 cp "$repo_root/scripts/quality_gate_secure_state.py" "$fixture/scripts/"
 cp "$repo_root/scripts/quality_gate_provenance.py" "$fixture/scripts/"
+cp "$repo_root/scripts/quality_gate_file_provenance.py" "$fixture/scripts/"
 cp "$repo_root/scripts/quality_gate_sandbox.py" "$fixture/scripts/"
 cp "$repo_root/scripts/quality_gate_tool_selection.py" "$fixture/scripts/"
 cp "$repo_root/scripts/quality_gate_host_attestation.py" "$fixture/scripts/"
